Bridge greenhouse location

+++ Data/Models/Greenhouse.cs
@@ -5,17 +5,24 @@ namespace Data.Models
 {
     public class Greenhouse
     {
-        [Key] public string GreenHouseId { get; set; }
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
-        [Required] public List<TemperatureMeasurement> TemperatureMesurments { get; set; }
-
-        [Required] public List<DioxideCarbonMeasurement> DioxideCarbonMeasurements { get; set; }
-
-        [Required] public List<HumidityMeasurement> HumidityMeasurements { get; set; }
-
-        [Required] public List<Threshold> Thresholds { get; set; }
-        [Required] public List<Pot> Pots { get; set; }
-        [Required] public List<SensorStatus> SensorStatuses { get; set; }
+        [Key]
+        public string GreenHouseId { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
+        [Required]
+        public List<TemperatureMeasurement> TemperatureMesurments { get; set; }
+        
+        [Required]
+        public List<DioxideCarbonMeasurement> DioxideCarbonMeasurements { get; set; }
+        
+        [Required]
+        public List<HumidityMeasurement> HumidityMeasurements { get; set; }
+        
+        [Required]
+        public List<Threshold> Thresholds { get; set; }
+        [Required]
+        public List<Pot> Pots { get; set; }
+        [Required]
+        public List<SensorStatus> SensorStatuses { get; set; }
     }
 }
